Use a common ThreadStartRoutinePrologue on all platforms.
